[0514] Update SIRNN SCI

+++ sci/LinearHE/utils-HE.cpp
@@ -27,22 +27,16 @@ using namespace sci;
 using namespace seal;
 using namespace seal::util;
 
-void generate_new_keys(
-        int SCI_party,
-        NetIO* io,
-        int slot_count,
-        shared_ptr<SEALContext> &context_,
-        Encryptor* &encryptor_,
-        Decryptor* &decryptor_,
-        Evaluator* &evaluator_,
-        BatchEncoder* &encoder_,
-        GaloisKeys* &gal_keys_,
-        Ciphertext* &zero_,
-        bool verbose)
+void generate_new_keys(int SCI_party, NetIO *io, int slot_count,
+                       shared_ptr<SEALContext> &context_,
+                       Encryptor *&encryptor_, Decryptor *&decryptor_,
+                       Evaluator *&evaluator_, BatchEncoder *&encoder_,
+                       GaloisKeys *&gal_keys_, Ciphertext *&zero_,
+                       bool verbose)
 {
     EncryptionParameters parms(scheme_type::BFV);
     parms.set_poly_modulus_degree(slot_count);
-    parms.set_coeff_modulus(CoeffModulus::Create(slot_count, {60,60,60,38}));
+    parms.set_coeff_modulus(CoeffModulus::Create(slot_count, {60, 60, 60, 38}));
     parms.set_plain_modulus(prime_mod);
     // auto context = SEALContext::Create(parms, true, sec_level_type::none);
     context_ = SEALContext::Create(parms, true, sec_level_type::none);
@@ -59,12 +53,12 @@ void generate_new_keys(
         pub_key.save(os);
         uint64_t pk_size = os.tellp();
         gal_keys_.save(os);
-        uint64_t gk_size = (uint64_t)os.tellp()-pk_size;
+        uint64_t gk_size = (uint64_t)os.tellp() - pk_size;
 
         string keys_ser = os.str();
         io->send_data(&pk_size, sizeof(uint64_t));
         io->send_data(&gk_size, sizeof(uint64_t));
-        io->send_data(keys_ser.c_str(), pk_size+gk_size);
+        io->send_data(keys_ser.c_str(), pk_size + gk_size);
 
 #ifdef HE_DEBUG
         stringstream os_sk;
@@ -83,21 +77,21 @@ void generate_new_keys(
         uint64_t gk_size;
         io->recv_data(&pk_size, sizeof(uint64_t));
         io->recv_data(&gk_size, sizeof(uint64_t));
-        char* key_share = new char[pk_size+gk_size];
-        io->recv_data(key_share, pk_size+gk_size);
+        char *key_share = new char[pk_size + gk_size];
+        io->recv_data(key_share, pk_size + gk_size);
         stringstream is;
         PublicKey pub_key;
         is.write(key_share, pk_size);
         pub_key.load(context_, is);
         gal_keys_ = new GaloisKeys();
-        is.write(key_share+pk_size, gk_size);
+        is.write(key_share + pk_size, gk_size);
         gal_keys_->load(context_, is);
         delete[] key_share;
 
 #ifdef HE_DEBUG
         uint64_t sk_size;
         io->recv_data(&sk_size, sizeof(uint64_t));
-        char* key_share_sk = new char[sk_size];
+        char *key_share_sk = new char[sk_size];
         io->recv_data(key_share_sk, sk_size);
         stringstream is_sk;
         SecretKey sec_key;
@@ -113,17 +107,13 @@ void generate_new_keys(
         zero_ = new Ciphertext;
         encryptor_->encrypt(tmp, *zero_);
     }
-    if (verbose) cout << "Keys Generated (slot_count: " << slot_count << ")" << endl;
+    if (verbose)
+        cout << "Keys Generated (slot_count: " << slot_count << ")" << endl;
 }
 
-void free_keys(
-        int SCI_party,
-        Encryptor* &encryptor_,
-        Decryptor* &decryptor_,
-        Evaluator* &evaluator_,
-        BatchEncoder* &encoder_,
-        GaloisKeys* &gal_keys_,
-        Ciphertext* &zero_)
+void free_keys(int SCI_party, Encryptor *&encryptor_, Decryptor *&decryptor_,
+               Evaluator *&evaluator_, BatchEncoder *&encoder_,
+               GaloisKeys *&gal_keys_, Ciphertext *&zero_)
 {
     delete encoder_;
     delete evaluator_;
@@ -142,42 +132,39 @@ void free_keys(
     }
 }
 
-void send_encrypted_vector(
-        NetIO* io,
-        vector<Ciphertext> &ct_vec)
+void send_encrypted_vector(NetIO *io, vector<Ciphertext> &ct_vec)
 {
     assert(ct_vec.size() > 0);
     stringstream os;
     uint64_t ct_size;
-    for(int ct = 0; ct < ct_vec.size(); ct++) {
+    for (size_t ct = 0; ct < ct_vec.size(); ct++)
+    {
         ct_vec[ct].save(os);
-        if(!ct) ct_size = os.tellp();
+        if (!ct)
+            ct_size = os.tellp();
     }
     string ct_ser = os.str();
     io->send_data(&ct_size, sizeof(uint64_t));
     io->send_data(ct_ser.c_str(), ct_ser.size());
 }
 
-void recv_encrypted_vector(
-        NetIO* io,
-        vector<Ciphertext> &ct_vec)
+void recv_encrypted_vector(NetIO *io, vector<Ciphertext> &ct_vec)
 {
     assert(ct_vec.size() > 0);
     stringstream is;
     uint64_t ct_size;
     io->recv_data(&ct_size, sizeof(uint64_t));
-    char* c_enc_result = new char[ct_size*ct_vec.size()];
-    io->recv_data(c_enc_result, ct_size*ct_vec.size());
-    for(int ct = 0; ct < ct_vec.size(); ct++) {
-        is.write(c_enc_result+ct_size*ct, ct_size);
+    char *c_enc_result = new char[ct_size * ct_vec.size()];
+    io->recv_data(c_enc_result, ct_size * ct_vec.size());
+    for (size_t ct = 0; ct < ct_vec.size(); ct++)
+    {
+        is.write(c_enc_result + ct_size * ct, ct_size);
         ct_vec[ct].unsafe_load(is);
     }
     delete[] c_enc_result;
 }
 
-void send_ciphertext(
-        NetIO* io,
-        Ciphertext &ct)
+void send_ciphertext(NetIO *io, Ciphertext &ct)
 {
     stringstream os;
     uint64_t ct_size;
@@ -188,14 +175,12 @@ void send_ciphertext(
     io->send_data(ct_ser.c_str(), ct_ser.size());
 }
 
-void recv_ciphertext(
-        NetIO* io,
-        Ciphertext &ct)
+void recv_ciphertext(NetIO *io, Ciphertext &ct)
 {
     stringstream is;
     uint64_t ct_size;
     io->recv_data(&ct_size, sizeof(uint64_t));
-    char* c_enc_result = new char[ct_size];
+    char *c_enc_result = new char[ct_size];
     io->recv_data(c_enc_result, ct_size);
     is.write(c_enc_result, ct_size);
     ct.unsafe_load(is);
@@ -203,10 +188,8 @@ void recv_ciphertext(
 }
 
 void set_poly_coeffs_uniform(
-        uint64_t *poly,
-        uint32_t bitlen,
-        shared_ptr<UniformRandomGenerator> random,
-        shared_ptr<const SEALContext::ContextData> &context_data)
+    uint64_t *poly, uint32_t bitlen, shared_ptr<UniformRandomGenerator> random,
+    shared_ptr<const SEALContext::ContextData> &context_data)
 {
     assert(bitlen < 128 && bitlen > 0);
     auto &parms = context_data->parms();
@@ -216,32 +199,39 @@ void set_poly_coeffs_uniform(
     uint64_t bitlen_mask = (1ULL << (bitlen % 64)) - 1;
 
     RandomToStandardAdapter engine(random);
-    for (size_t i = 0; i < coeff_count; i++) {
-        if (bitlen < 64) {
+    for (size_t i = 0; i < coeff_count; i++)
+    {
+        if (bitlen < 64)
+        {
             uint64_t noise = (uint64_t(engine()) << 32) | engine();
             noise &= bitlen_mask;
-            for (size_t j = 0; j < coeff_mod_count; j++) {
-                poly[i + (j * coeff_count)] = barrett_reduce_63(noise, coeff_modulus[j]);
+            for (size_t j = 0; j < coeff_mod_count; j++)
+            {
+                poly[i + (j * coeff_count)] =
+                    barrett_reduce_63(noise, coeff_modulus[j]);
             }
-        } else {
+        }
+        else
+        {
             uint64_t noise[2]; // LSB || MSB
-            for(int j = 0; j < 2; j++){
+            for (int j = 0; j < 2; j++)
+            {
                 noise[0] = (uint64_t(engine()) << 32) | engine();
                 noise[1] = (uint64_t(engine()) << 32) | engine();
             }
             noise[1] &= bitlen_mask;
-            for (size_t j = 0; j < coeff_mod_count; j++) {
-                poly[i + (j * coeff_count)] = barrett_reduce_128(noise, coeff_modulus[j]);
+            for (size_t j = 0; j < coeff_mod_count; j++)
+            {
+                poly[i + (j * coeff_count)] =
+                    barrett_reduce_128(noise, coeff_modulus[j]);
             }
         }
     }
 }
 
-void flood_ciphertext(
-        Ciphertext &ct,
-        shared_ptr<const SEALContext::ContextData> &context_data,
-        uint32_t noise_len,
-        MemoryPoolHandle pool)
+void flood_ciphertext(Ciphertext &ct,
+                      shared_ptr<const SEALContext::ContextData> &context_data,
+                      uint32_t noise_len, MemoryPoolHandle pool)
 {
 
     auto &parms = context_data->parms();
@@ -253,16 +243,18 @@ void flood_ciphertext(
     shared_ptr<UniformRandomGenerator> random(parms.random_generator()->create());
 
     set_poly_coeffs_uniform(noise.get(), noise_len, random, context_data);
-    for (size_t i = 0; i < coeff_mod_count; i++) {
-        add_poly_poly_coeffmod(noise.get() + (i * coeff_count), 
-            ct.data() + (i * coeff_count), coeff_count, 
-            coeff_modulus[i], ct.data() + (i * coeff_count));
+    for (size_t i = 0; i < coeff_mod_count; i++)
+    {
+        add_poly_poly_coeffmod(noise.get() + (i * coeff_count),
+                               ct.data() + (i * coeff_count), coeff_count,
+                               coeff_modulus[i], ct.data() + (i * coeff_count));
     }
 
     set_poly_coeffs_uniform(noise.get(), noise_len, random, context_data);
-    for (size_t i = 0; i < coeff_mod_count; i++) {
-        add_poly_poly_coeffmod(noise.get() + (i * coeff_count), 
-            ct.data(1) + (i * coeff_count), coeff_count, 
-            coeff_modulus[i], ct.data(1) + (i * coeff_count));
+    for (size_t i = 0; i < coeff_mod_count; i++)
+    {
+        add_poly_poly_coeffmod(noise.get() + (i * coeff_count),
+                               ct.data(1) + (i * coeff_count), coeff_count,
+                               coeff_modulus[i], ct.data(1) + (i * coeff_count));
     }
 }
